refactor: SIMT-92 스케줄 확인 및 삭제 에외처리
