+++ feature/feature-home/src/main/java/com/comit/feature_home/navigation/HomeNavigation.kt
@@ -13,7 +13,7 @@ import com.comit.feature_home.screen.SalaryWebViewScreen
 import com.comit.feature_home.screen.alarm.AlarmScreen
 import com.comit.feature_home.screen.schedule.ShowScheduleScreen
 import com.comit.feature_home.screen.schedule.WriteScheduleScreen
-import com.comit.feature_home.vm.WriteClosedDayScreen
+import com.comit.feature_home.screen.WriteClosedDayScreen
 import com.comit.navigator.SimTongScreen
 
 fun NavGraphBuilder.homeNavigation(
